Reduced the amount of points to check through. All objects near the 2 players are being disregarded as this was where the most false positives were being found

+++ SquashBallTracker/TrajectoryPreditor.hpp
@@ -14,8 +14,8 @@ namespace tpd {
 
 	private:
 
-		static std::vector<float> polynomialPath(std::vector<cv::Rect>);
-		static float calcDeterminant(std::vector<std::vector<float>> );
+		static std::vector<double> polynomialPath(std::vector<cv::Rect>);
+		static long long calcDeterminant(std::vector<std::vector<unsigned long long>>);
 	};
